Changed the api to be full asynchronous using spring webflux

+++ SERVER/WEB_API/src/main/java/isel/ps/employbox/services/CompanyService.java
@@ -1,36 +1,78 @@
 package isel.ps.employbox.services;
 
+import isel.ps.employbox.ErrorMessages;
+import isel.ps.employbox.exceptions.BadRequestException;
 import isel.ps.employbox.exceptions.ResourceNotFoundException;
+import isel.ps.employbox.exceptions.UnauthorizedException;
+import isel.ps.employbox.model.entities.Account;
 import isel.ps.employbox.model.entities.Company;
+import isel.ps.employbox.model.entities.Role;
+import javafx.util.Pair;
 import org.github.isel.rapper.DataRepository;
 import org.springframework.stereotype.Service;
+import reactor.core.publisher.Mono;
 
-import java.util.Optional;
+import java.util.Collection;
+import java.util.concurrent.CompletableFuture;
 import java.util.stream.Stream;
-import java.util.stream.StreamSupport;
 
 @Service
 public class CompanyService {
+    private DataRepository<Account, Long> accountRepo;
     private DataRepository<Company, Long> companyRepo;
 
-    public CompanyService(DataRepository<Company, Long> companyRepo){ this.companyRepo = companyRepo; }
+    public CompanyService( DataRepository<Account, Long> accountRepo, DataRepository<Company, Long> companyRepo){
+        this.accountRepo = accountRepo;
+        this.companyRepo = companyRepo; }
 
-    public Stream<Company> getCompanies() {
-        return StreamSupport.stream(companyRepo.findAll().join().spliterator(),false);
+    public CompletableFuture<Stream<Company>> getCompanies() {
+        return companyRepo.findAll().thenApply(Collection::stream);
     }
 
-    public Company getCompany(long cid) {
-        Optional<Company> ocomp = companyRepo.findById(cid).join();
-        if(!ocomp.isPresent())
-            throw new ResourceNotFoundException("Company doesnt exist");
-        return ocomp.get();
+    public CompletableFuture<Company> getCompany(long cid) {
+        return companyRepo.findById(cid)
+                .thenApply(
+                        optionalCompany -> optionalCompany.orElseThrow(() -> new ResourceNotFoundException(ErrorMessages.resourceNotfound_company)));
     }
 
-    public void createCompany(Company company, String name) { companyRepo.create(company); }
-
-    public void deleteCompany(long cid, String email) {
-        companyRepo.deleteById(cid);
+    public CompletableFuture<Company> createCompany(Company company, String email) {
+        return accountRepo.findWhere(new Pair<>("email", email))
+                .thenCompose(list -> {
+                            if (list.get(0).getRole() != Role.ADMINISTRATOR)
+                                throw new UnauthorizedException(ErrorMessages.unAuthorized);
+                            return companyRepo.create(company);
+                        }
+                ).thenApply(res -> {
+                    if (!res) throw new BadRequestException(ErrorMessages.badRequest_ItemCreation);
+                    return company;
+                });
     }
 
-    public void updateCompany(Company company, String name) { companyRepo.update(company); }
+    public Mono<Void> deleteCompany(long cid, String email) {
+        return Mono.fromFuture(
+                accountRepo.findWhere(new Pair<>("email", email))
+                        .thenCompose(list -> {
+                                    if (list.get(0).getRole() != Role.ADMINISTRATOR)
+                                        throw new UnauthorizedException(ErrorMessages.unAuthorized);
+                                    return companyRepo.deleteById(cid);
+                                }
+                        ).thenAccept(res -> {
+                    if (!res) throw new BadRequestException(ErrorMessages.badRequest_ItemDeletion);
+                })
+        );
+    }
+
+    public Mono<Void> updateCompany(Company company, String email) {
+        return Mono.fromFuture(
+                accountRepo.findWhere(new Pair<>("email", email))
+                        .thenCompose(list -> {
+                                    if (list.get(0).getRole() != Role.ADMINISTRATOR)
+                                        throw new UnauthorizedException(ErrorMessages.unAuthorized);
+                                    return companyRepo.update(company);
+                                }
+                        ).thenAccept(res -> {
+                    if (!res) throw new BadRequestException(ErrorMessages.badRequest_ItemUpdate);
+                })
+        );
+    }
 }
